requirements.txt
streamlit
python-frontmatter
pyyaml
python-slugify
markdown
numpy
rank-bm25 